"interface" -> "UI" nomenclature change
In hindsight, it was a mistake to use "interface" as my nomenclature for UI-related matters. It gets messy when speaking about interfaces in the programming sense. I'll use UI instead. I try to avoid changes like this because they can harm a team's familiarity with the code base and they obfuscate the git history of renamed classes. This feels like a worthwhile time though.